add Manipulation
have problem about bfs

+++ Cell/Cell.cpp
@@ -3,7 +3,7 @@
 void Cell::SetIndex(int x, int y) {
     idX = x, idY = y;
 }
-void Cell::SetPosition(float x, float y) {
+void Cell::SetPosition(int x, int y) {
     shape.setPosition(x, y);
 }
 RectangleShape& Cell::GetRectangleShape() {
